make triton group_norm faster

+++ src/sfast/triton/ops/activation.py
@@ -9,9 +9,9 @@ def identity(x):
 
 @triton.jit
 def silu(x):
-    return x * tl.sigmoid(x)
+    return x * tl.sigmoid(x.to(tl.float32)).to(x.dtype)
 
 
 @triton.jit
 def relu(x):
-    return tl.max(x, 0)
+    return tl.max(x, 0.0)
